activated info for unnecessary casts and instanceof checks and fixed all existing ones

+++ game-server/data/scripts/system/handlers/instance/IlluminaryObeliskInstance.java
@@ -274,7 +274,7 @@ public class IlluminaryObeliskInstance extends GeneralInstanceHandler {
 				TeleportService2.teleportTo(player, mapId, instanceId, 265.45142f, 264.52875f, 455.1256f, (byte) 75);
 				break;
 			case 702009:
-				TeleportService2.teleportTo(player, mapId, instanceId, npc.getX(), npc.getY(), npc.getZ(), (byte) npc.getHeading(), TeleportAnimation.FADE_OUT_BEAM);
+				TeleportService2.teleportTo(player, mapId, instanceId, npc.getX(), npc.getY(), npc.getZ(), npc.getHeading(), TeleportAnimation.FADE_OUT_BEAM);
 				SkillEngine.getInstance().applyEffectDirectly(21511, player, player, 0);
 				npc.getController().onDelete();
 				break;
@@ -296,7 +296,7 @@ public class IlluminaryObeliskInstance extends GeneralInstanceHandler {
 	@Override
 	public void onEndEffect(Creature effector, Creature effected, int skillId) {
 		if (skillId == 21511)
-			spawn(702009, effected.getX(), effected.getY(), effected.getZ(), (byte) effected.getHeading());
+			spawn(702009, effected.getX(), effected.getY(), effected.getZ(), effected.getHeading());
 	}
 	
 	@Override
